removed utils benchmark
